Defect & Step Length
Added point defect and modified step length

+++ general/include/ECHARM_intrp.hh
@@ -20,7 +20,7 @@ public:
 
     double GetXinUnitPeriod(double,double&);
     double GetXinUnitOne(double,double&);
-    void GetIndexes(int*,int&);
+    void GetIndexes(long*,int&);
 };
 
 #endif
